refactor: Socket global event handler module
- move integration of hooks, redux actions, `socketEventManager` and event handler functions from `useSocial.js` to `useGlobalEvents.js`.
- add social events with their matching handlers to `subscriptions` array, and `useEffect` dependency array in `useGlobalEvents.js`.
- move logic for subscribing to `exam finished` socket event from `Header.jsx`. to `useGlobalEvents.js`.
- remove integration of `finishExam` redux action, and `socketEventManager` in `Header.jsx`.
- integrate `useGlobalEvents` hook in `Header.jsx`.
- delete the redundant custom hook `useSocial.js`.
